+++ docs/code-review/step1-01.md
@@ -48,11 +48,70 @@
 학습은 분명 중요하지만 그 핵심 과정으로 리뷰/피드백이 존재하는 이상,  
 효율적인 리뷰를 위해 더 많은 시간을 배분하고 일정을 조율해야겠다 생각했습니다.
 
-### 📝 Feedback 0
+### 📝 Feedback 02
+
+> 인터페이스에 `@Repository`를 붙이는 게 의미가 있을까요? 해당 어노테이션이 왜 있는건지 학습해보셔요
+
+### 💬 Apply 02
+
+#### 인터페이스에 붙은 어노테이션 
+
+명백한 검증 누락입니다. 이전 미션에서 피드백 받은 내용을 반영하지 못했네요.  
+
+> 커밋 전 diff 확인과 자기 작업의 실제 결과를 검증하는 습관이 보강되면 완성도가 크게 올라갈 것입니다.
+
+작업 단위를 잘게 쪼개고 커밋 전에 확인할 수 있도록,  
+또 그럴 여유를 충분히 가지기 위해  
+더 세부적인 일정 계획을 수립할 수 있도록 개선하겠습니다. 짚어주셔서 감사합니다!  
+
+#### `@Repository` 어노테이션
+
+우선 직관적으로 와닿은 Spring Repository 의 차이점은 `throws SQLException` 의 배제였습니다.  
+이전 미션에선 Repository / DAO 의 모든 Connection 메서드에 throws 구문을 붙여야 했지만  
+그것이 생략되었다는 것은 Spring 이 어떤 방식으로나마 예외를 처리(not Solve)해 주기 때문이라 생각했어요.  
+
+어노테이션의 Javadoc 내용을 활용해 학습해봤어요.
+
+`번역`  
+
+- 이 어노테이션이 붙은 클래스는 도메인 주도 설계(DDD)에서 정의한  
+  "저장, 조회, 검색 행위를 캡슐화하여 마치 객체 컬렉션처럼 동작하는 메커니즘"인 'Repository' 임을 나타냅니다.
+- [X] 전통적인 J2EE 패턴인 'DAO(Data Access Object)'를 구현하는 팀들도  
+  이 어노테이션을 사용할 수 있지만, 적용하기 전에  
+  DAO와 DDD 스타일 Repository의 차이를 이해하는 것이 좋습니다.
+- 이 어노테이션이 붙은 클래스는 PersistenceExceptionTranslationPostProcessor와 함께 사용될 때,  
+  하부 데이터베이스의 원시 예외를 스프링의 범용적인 DataAccessException으로  
+  자동 변환(Translation)해 주는 자격을 얻습니다.
+- 또한, 이 어노테이션은 @Component의 특수한 형태이므로 클래스패스 스캔을 통해  
+  스프링 컨테이너에 자동으로 빈(Bean)으로 등록됩니다.
+
+생각했던 대로 `데이터베이스의 원시 예외를 범용직인 예외로 변환` 한다는 내용이 포함되어 있었습니다.  
+또한,  
+데이터베이스의 원시 예외는 사용하는 기술에 따라 다르기에 그것을 직접 던진다면  
+서비스는 예외 처리에 데이터베이스의 종류에 의존하게 되고, 그것을 방지하기 위해  
+`@Repository` 어노테이션이 예외를 변환해서 던져 주는 것임을 확인할 수 있었어요.  
+
+이런 계층간 의존/종속 제거에 [앞선 답변]() 내용이 포함된다 생각했어요.  
+예외 처리 측면에서 예외를 포장해서 던짐으로써 하부 데이터베이스의 종류에 대한 종속을 제거하듯  
+기능 구현 측면에서 저장소를 인터페이스로 추상화함으로써 동일한 종속 제거가 가능한 것이 아닐까요?  
+
+#### 정리 
+
+일단 여전히 셀프 검증과 디테일 체크가 부족함을 확인할 수 있었어요.  
+처음에 클래스를 생성하며 자연스럽게 어노테이션을 붙이고  
+인터페이스로 변경 -> 구현체 구현 과정에서 내내 까먹고 있었네요..  
+
+diff 나 IDE 기능을 활용해서 검증하는 습관을 들여보도록 하겠습니다. 감사합니다!!
+
+실제 구현 과정에서 북스터디나 학습에서 배운 내용이 부분부분 튀어나오니  
+소소하게 보람도 있고 바로바로 나름의 답변을 낼 수 있다는 것이 매우 즐겁습니다.  
+틀린 부분이나 더 공부할 부분 팍팍 지적해주시면 열심히 고치고 배우겠습니다!  
+
+### 📝 Feedback 03
 
 
 
-### 💬 Apply 0
+### 💬 Apply 03
