+++ src/main/java/roomescape/controller/ReservationRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.controller;
 
 public record ReservationRequest(String name, String date, long timeId) {
 }
