+++ src/main/java/roomescape/controller/ReservationResponse.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.controller;
 
 public record ReservationResponse(long id, String name, String date, ReservationTimeResponse time) {
 }
