+++ src/main/java/roomescape/controller/ReservationTimeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.controller;
 
 public record ReservationTimeResponse(long id, String startAt) {
 }
