+++ src/main/java/roomescape/repository/JdbcReservationDao.java
@@ -8,7 +8,7 @@ import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
-import roomescape.dto.ReservationRequest;
+import roomescape.controller.ReservationRequest;
 
 @Repository
 public class JdbcReservationDao implements ReservationDao {
