+++ src/main/java/roomescape/repository/ReservationTimeDao.java
@@ -2,7 +2,7 @@ package roomescape.repository;
 
 import java.util.List;
 import roomescape.domain.ReservationTime;
-import roomescape.dto.ReservationTimeRequest;
+import roomescape.controller.ReservationTimeRequest;
 
 public interface ReservationTimeDao {
